+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -11,7 +11,7 @@ CHANGELOG
  * Added PHPUnit constraints
  * Added `MessageDataCollector`
  * Added `MessageEvents` and `MessageLoggerListener` to allow collecting sent emails
- * [BC BREAK] `TransportInterface` has a new `getName()` method
+ * [BC BREAK] `TransportInterface` has a new `__toString()` method
  * [BC BREAK] Classes `AbstractApiTransport` and `AbstractHttpTransport` moved under `Transport` sub-namespace.
  * [BC BREAK] Transports depend on `Symfony\Contracts\EventDispatcher\EventDispatcherInterface`
    instead of `Symfony\Component\EventDispatcher\EventDispatcherInterface`.